test: split GitHub workflow policy contracts (#2212)
